move wallet to own file
